Remove interface now that there is only one implementation
